Added JQuery and template for logging module
Added JQuery to the example project.

Updated git ignore to avoid commiting cljsbuild files in the example
projects.

Added a base to build a logging feature.

Fixed the keyboaard handler so it registers itself, but still won't work
as there is no way currently to access the feature specific state.